Issue #2: Incorrect header set when calling renderStatus(x)
There is a hard coded 200 status code in views.js that overrides any
useful status code sent when calling renderStatus to render an error
template.

By replacing the 200 with res.statusCode, the template now renders and
transmits with an appropriate HTTP status code.

I have added a test but it will need to be verified that it is a valid
test as I am new to node and node's assert module.